Revert "Bugfix/leaves"

+++ app/models/payroll.py
@@ -35,7 +35,7 @@ class PayrollLeaveAdjustment(Base):
     deduct = Column(Boolean, nullable=False, default=True)
     # Snapshot of unpaid working-days for this leave within the run's month.
     # Null on legacy rows → fall back to (days_in_month if deduct else 0).
-    unpaid_days = Column(Float, nullable=True)
+    unpaid_days = Column(Integer, nullable=True)
     created_at = Column(TIMESTAMP, server_default=func.now())
 
     __table_args__ = (UniqueConstraint("payroll_run_id", "leave_id", name="uq_payroll_adj_run_leave"),)
